% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allCommonDistinctiveStructures.R
\name{allCommonDistinctive}
\alias{allCommonDistinctive}
\title{allCommonDistinctive: creates all common and distinctive structures for a component weight matrix}
\usage{
allCommonDistinctive(
  vars,
  ncomp,
  allPermutations = TRUE,
  filterZeroSegments = TRUE
)
}
\arguments{
\item{vars}{A vector indicating the number of variables per data block. Example, c(10, 20, 5), there the first 10 variables belong the first block, the next 20 variables belong to the second block, the next 5 variables belong to the third block}

\item{ncomp}{The number of components that are of interest}

\item{allPermutations}{A boolean with default \code{TRUE}. If set to \code{TRUE}, all permutations of common and distinctive will be taken into account (a certain component can occur in all positions). If set to \code{FALSE} all permutations will not be taken into account i.e. a certain component can only occur once. If the intent is to use multiple starts use \code{FALSE} else, use \code{TRUE}}

\item{filterZeroSegments}{A boolean with default \code{TRUE}. If set to \code{TRUE} components were all blocks are set to zero are not possible reducing the number of combinations.}
}
\value{
A list with all common and distinctive structures
}
\description{
A function that generates all common and distinctive structures given the number of variables per block and the number of components
}
\examples{

allCommonDistinctive(vars = c(10, 5, 5), ncomp = 3, allPermutations = TRUE, filterZeroSegments = TRUE)
allCommonDistinctive(vars = c(10, 5, 5), ncomp = 3, allPermutations = FALSE, filterZeroSegments = FALSE)
 
}
