% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVforPCAwithSparseWeights.R
\name{CVforPCAwithSparseWeights}
\alias{CVforPCAwithSparseWeights}
\title{Cross-validation with the EigenVector method (CV) for sparse weight based PCA methods}
\usage{
CVforPCAwithSparseWeights(X, nrFolds, FUN, ...)
}
\arguments{
\item{X}{A data matrix of class 'matrix'}

\item{FUN}{A pointer to a function (i.e. the function name with no brackets) that performs sparse weight based PCA. it should return a list containing a matrix object called "W" that contains the component weights}

\item{...}{specify all the arguments the function in FUN needs}
}
\value{
The following items in a list \cr
\code{MSPE} The MSPE given the tuning parameters 
\code{MSPEstdError} The standard error of the MSPE 
\code{nNonZeroCoef} The number of non-zero coefficients in the model
}
\description{
A function that returns the mean squared prediction error (MSPE) given a set of tuning parameters and a sparse weight based PCA method
}
\examples{
 
}
