% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISforPCAwithSparseWeights.R
\name{ISforPCAwithSparseWeights}
\alias{ISforPCAwithSparseWeights}
\title{Index of sparseness (IS) for sparse weight based PCA methods}
\usage{
ISforPCAwithSparseWeights(X, ncomp, FUN, ...)
}
\arguments{
\item{X}{A data matrix of class 'matrix'}

\item{ncomp}{An integer specifying the number of components}

\item{FUN}{A pointer to a function (i.e. the function name with no brackets) that performs sparse weight based PCA. it should return a list containing a matrix object called "W" that contains the component weights}

\item{...}{specify all the arguments the function in FUN needs}
}
\value{
The following items in a list \cr
\code{IS} The IS given the tuning parameters 
\code{nNonZeroCoef} The number of non-zero coefficients in the model
}
\description{
A function that returns the IS given a set of tuning parameters and a sparse weight based PCA method
}
\examples{
 
}
