% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeData.R
\name{makeDat}
\alias{makeDat}
\title{makeDat: creates simulated data based on a sparse PCA model}
\usage{
makeDat(n, comdis, variances)
}
\arguments{
\item{n}{The number of objects the data should have}

\item{comdis}{A \code{matrix} specifying the zero structure of \eqn{W}, the data will have Q = \code{ncol(comdis)} "important components" and J = \code{nrow(comdis)} variables}

\item{variances}{specifying the variances of the J components these are the J eigenvalues of \eqn{X^T X}}

\item{ncomp}{The number of components that are of interest}
}
\value{
A list with the following items: \cr
\code{X} A data matrix generated from MASS::mvrnorm() with a zero mean structure and Sigma = P \%*\% diag(variances) \%*\% t(P), empirical is set \code{FALSE} \cr
\code{P} A matrix of dimension J x J, with the loadings/weights the first Q columns have the sparsity structure specified in \code{comdis}, the other Q-J columns are non-sparse. \cr
\code{Sigma} The covariance matrix that is used to generate the data from Sigma \code{= P \%*\% diag(variances) \%*\% t(P)} \cr
In case of failure the function returns \code{NA}. The function can fail if the \code{comdis} structure specified in P is not possible, i.e. linear dependency
}
\description{
A function that generates a realistic dataset based on a sparse PCA model \eqn{X = XWP^T}, with a sparse \eqn{W} and with W = P and \eqn{W^T W = I}
}
\examples{

ncomp <- 3 
J <- 30
comdis <- matrix(1, J, ncomp)
comdis <- sparsify(comdis, 0.7) #set 70\% of the 1's to zero
variances <- makeVariance(varianceOfComps = c(100, 80, 70), J = J, error = 0.05) #create realistic eigenvalues
dat <- makeDat(n = 100, comdis = comdis, variances = variances)
 
}
