% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeData.R
\name{makeVariance}
\alias{makeVariance}
\title{makeVariance: Create variances for the components in data generation}
\usage{
makeVariance(varianceOfComps, J, error)
}
\arguments{
\item{J}{The total number of components (the number of variables)}

\item{error}{The total variance of J-Q components compared to the variance of the Q components}

\item{varianceOfcomp}{The variances of the first Q components of interest}
}
\value{
A vector with J variances (eigenvalues) to simulate data from.
}
\description{
A function that creates realistic variances for J components, the variances are the J eigenvalues of a I x J data matrix X. The variances of the first Q components are set by the user, the remaining J - Q variances 
are decreasing on a reversed logscale. The error
}
\examples{

variances <- makeVariance(c(120, 100, 80), J = 50, error = 0.05) 
plot(variances) #realistic eigenvalues of a data matrix X

}
