\name{sparseWeightBasedPCA-package}
\alias{sparseWeightBasedPCA-package}
\alias{sparseWeightBasedPCA}
\docType{package}
\title{
\packageTitle{sparseWeightBasedPCA}
}
\description{
\packageDescription{sparseWeightBasedPCA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sparseWeightBasedPCA}
\packageIndices{sparseWeightBasedPCA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sparseWeightBasedPCA}

Maintainer: \packageMaintainer{sparseWeightBasedPCA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
