% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeData.R
\name{sparsify}
\alias{sparsify}
\title{Sparsify: shoots a percentage of zero's in non-zero elements of columns of a matrix}
\usage{
sparsify(comdis, sparsity)
}
\arguments{
\item{comdis}{A data matrix of class 'matrix', with a common or distinctive structure. (i.e. zero and ones in prespecified locations)}

\item{sparsity}{The percentage of non-zero elements in the columns of comdis that should be set to zero. This is a value between zero and 1}
}
\value{
A matrix with a percantage of the non-zero elments in comdis set to zero
}
\description{
Sparsify: set at random a percentage the non-zero elements of columns of a matrix to zero. This can be used to create a zerp structure for the component weights/loadings from which to simmulate data.
}
\examples{

comdis <- matrix(1, 40, 3) 
sparsify(comdis, 0.7) #set 70\% of the non-zero elements in comdis to zero

}
