% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccpca}
\alias{ccpca}
\title{This function performs PCA with cardinality constraints on the component weights.
Note: This function is fully written in C++, therefore no default parameter arguments are given, please check the example below for sensible default values.}
\usage{
ccpca(
  X,
  ncomp,
  nzeros,
  itr,
  Wstart,
  nStarts = 1L,
  tol = 1e-07,
  printLoss = TRUE
)
}
\arguments{
\item{X}{A data matrix of class 'matrix'}

\item{ncomp}{The number of components to estimate (an integer)}

\item{nzeros}{A vector of length ncomp containing the number of desired zeros in the columns of the component weight matrix W}

\item{itr}{The maximum number of iterations (an integer)}

\item{Wstart}{A matrix of ncomp columns and nrow(X) rows with starting values for the component weight matrix W, if Wstart only contains zeros, a warm start is used: the first ncomp right singular vectors of X}

\item{nStarts}{The number of random starts the analysis should perform. The first start will be performed with the values given by Wstart. The consecutive starts will be Wstart plus a matrix with random uniform values times the current start number (the first start has index zero). The default value is 1.}

\item{tol}{The convergence is determined by comparing the loss function value after each iteration, if the difference is smaller than tol, the analysis is converged. The default value is 10e-8}

\item{printLoss}{A boolean: TRUE will print the lossfunction value each 10th iteration.}
}
\value{
A list containing: \cr
\code{W} A matrix containing the component weights \cr
\code{P} A matrix containing the loadings \cr
\code{loss} A numeric variable ccontaining the minumn loss function value of all the nStarts starts \cr
\code{converged} A boolean containing \code{TRUE} if converged \code{FALSE} if not converged.
}
\description{
This function performs PCA with cardinality constraints on the component weights.
Note: This function is fully written in C++, therefore no default parameter arguments are given, please check the example below for sensible default values.
}
\examples{

I <- 100
J <- 50 
ncomp <- 3
X <- matrix(rnorm(I*J), I, J)

ccpca(X = X, ncomp = ncomp,  nzeros = c(10, 20, 30), itr = 100000, 
     Wstart = matrix(0, J, ncomp), nStarts = 1, tol = 10^-8, printLoss = TRUE)
}
